create save function to save file
